requirements.txt
flask == 1.1.1
pillow == 6.1.0